+++ eng/tools/sdk-suppressions/package.json
@@ -15,8 +15,6 @@
     "node": ">= 18.0.0"
   },
   "dependencies": {
-    "@octokit/rest": "18.0.3",
-    "@octokit/types": "^5.4.0",
     "ajv": "^6.12.6",
     "js-yaml": "^3.14.0",
     "lodash": "^4.17.20"
